requirements.txt
streamlit==1.28.2
deepface==0.0.81
pillow==10.1.0